+++ src/main/java/br/com/test/miniautorizador/service/CardService.java
@@ -1,5 +1,7 @@
 package br.com.test.miniautorizador.service;
 
+import java.math.BigDecimal;
+
 import br.com.test.miniautorizador.domain.CreateCardDTO;
 import br.com.test.miniautorizador.domain.ResponseCardDTO;
 import br.com.test.miniautorizador.domain.TransactionDTO;
@@ -8,7 +10,7 @@ public interface CardService {
 
 	ResponseCardDTO create(CreateCardDTO cardDTO) ;
 
-	Double getBalance(String cardNumber);
+	BigDecimal getBalance(String cardNumber);
 
 	void doTransaction(TransactionDTO transactionDTO);
 }
